uboot/kernel: Use the right version of mkimage
The "make uImage" step in the kernel build process runs mkimage - but
given the u-boot build process doesn't put mkimage in the PATH and
the kernel just calls mkimage without any path prefix, that results
in the system wide version of mkimage being called - that mkimage
could be from a very different version of u-boot, and that's kind of
asking for trouble.
Also fixes the build if no system wide mkimage is available.

Fixes https://bugs.launchpad.net/linaro-android/+bug/877320

Signed-off-by: Bernhard Rosenkraenzer <Bernhard.Rosenkranzer@linaro.org>