tasks/bootloader.mk: Add support for additional/custom bootloaders
This adds a mechanism for building additional versions of U-Boot by
defining their configs and names in the variable UBOOT_FLAVOURS. This
will be used by Vexpress to provide a single Android image which will
work with A5, A9 and A15 CPUs.

This change also provides a hook for boards to provide customised rules
for building bootloaders (CUSTOM_BOOTLOADER_MAKEFILE).

It is envisaged that this new task makefile and hooks will be used to
add support for using a UEFI bootloader and a means of tidying up the
way U-Boot is built in uboot.mk and bootfiles.mk.

Change-Id: I8bdddab60e429a8bdf4f022b752b524a980afc5e
Signed-off-by: Jon Medhurst <tixy@linaro.org>