tasks/kernel.mk: Conditionally build out-of-tree gator module
For the linux-linaro tree the Gator module will be in the kernel
tree. Therefore make the rule for building the out-of-tree module
conditional on the presence of the external source code.

This change doesn't attempt to handle in-tree gator modules built with
TARGET_GATOR_WITH_MALI_SUPPORT. We will leave that until we have such a
case as the linux-linaro single tree may lead to some consolidation of
Mali drivers.

Change-Id: I03facdc6cae768409887f9a6cb46b28e2ae96fec
Signed-off-by: Jon Medhurst <tixy@linaro.org>