kernel.mk: Enable use of kernel config fragments
If a board specifies KERNEL_CONFIG with more than one whitespace
separated value, then instead of been considered a defconfig to be made,
it is treated as a list of config fragments which need merging.

E.g. for vexpress we will now be able to specify:

KERNEL_CONFIG := linaro/configs/linaro-base.conf \
                 linaro/configs/android.conf \
                 linaro/configs/vexpress.conf

Change-Id: I8ede61ee9334b1f3b9e5b99291e529a46fb22214
Signed-off-by: Jon Medhurst <tixy@linaro.org>