bootloader.mk: Add helpers for building RTSM boot-wrappers
To successfully build the RTSM boot-wrapper requires careful selection
of an appropriate toolchain. This patch adds a helper macro,
MAKE_RTSM_BOOTWRAPPER, which consolidates this complexity in one place.

Below is an example of its use, in this case to construct a rule
to build the semi-hosing boot-wrapper.

$(BOOTLOADER_OUT)/linux-system-semi.axf : $(ACP) FORCE_BOOTLOADER_REMAKE
	cd $(TOP)/boot-wrapper && $(MAKE_RTSM_BOOTWRAPPER) linux-system-semi.axf
	@mkdir -p $(dir $@)
	$(ACP) -fpt $(TOP)/boot-wrapper/linux-system-semi.axf $@

BOOTLOADER_TARGETS += $(BOOTLOADER_OUT)/linux-system-semi.axf

Change-Id: I7173a794564097fc6597d1b391a61a425db4090f
Signed-off-by: Jon Medhurst <tixy@linaro.org>