common: Copy build specific HOWTO
This commit is the first instance of the LINARO_BUILD_SPEC
variable. This var is meant to be set to something like
panda-jb-gcc47-tilt-stable-blob. This feature is generally applicable
other places and may be able to replace TARGET_PRODUCT.

The Makefile changes cover invalid values of this and empty values. In
both of these cases a default.xml will be used.

Change-Id: Ie1fff4507f0fdaf0474439ab432d207b8366420d
Signed-off-by: Zach Pfeffer <zach.pfeffer@linaro.org>