kernel.mk: Only build uImage when using U-Boot
Change the default KERNEL_TARGET to zImage and only use uImage when
U-Boot is specified as the bootloader. Additionally, only set the
KERNEL_TARGET if it is not already specified, this will let devices
override the default.

These changes will help support moving to building zImages rather than
uImages which are more friendly for most bootloaders. Note, U-Boot can
also support zImage and will likely need to do so soon as uImage is
broken on multiplatform kernels due to the load address in the header
never being set.

Change-Id: I58eadf2c84452d5f968fd68e548c49c3dacea21a
Signed-off-by: Jon Medhurst <tixy@linaro.org>