bootfiles.mk: make PRE_BOOT_FILES dependency of boottarball
If we do not make this, the prebuilt boot files might get copied
after boot.tar.bz2 generated.
it can be reproduced when specify -j1 for the make command.

For using it, we only need to specify the list of prebuilt boot files
as value of PRE_BOOT_FILES like following:

    PRE_BOOT_FILES := bl1.bin bl2.bin bl31.bin

Change-Id: I1eb6d15b4a46d2f94fe1eff15e776515afe1d4fc
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>